move views to dedicated package, update acls
