+++ samples/spring-valueentity-customer-registry/src/main/java/com/example/Main.java
@@ -14,7 +14,7 @@ import org.springframework.context.annotation.Import;
 // Kalix service, but disallow access from the internet. This can be overridden explicitly
 // per component or method using annotations.
 // Documentation at https://docs.kalix.io/services/using-acls.html
-@Acl(allow = @Acl.Matcher(service = "*"))
+@Acl(allow = @Acl.Matcher(principal = Acl.Principal.INTERNET))
 public class Main {
 
   private static final Logger logger = LoggerFactory.getLogger(Main.class);
